pyproject.toml
[project]
name = "github-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.3",
    "python-dotenv>=1.1.1",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-core>=0.3.72",
    "asyncio>=3.4.3",
    "langgraph-supervisor>=0.0.29",
    "langchain_google_genai",
    "langchain-google-vertexai>=2.0.28",
]
